+++ base/src/test/java/org/aya/test/ThrowingReporter.java
@@ -5,6 +5,7 @@ package org.aya.test;
 import org.aya.api.distill.DistillerOptions;
 import org.aya.api.error.Problem;
 import org.aya.api.error.Reporter;
+import org.aya.cli.single.CliReporter;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.TestOnly;
 import org.junit.jupiter.api.Assertions;
@@ -14,7 +15,7 @@ public record ThrowingReporter() implements Reporter {
   public static final @NotNull ThrowingReporter INSTANCE = new ThrowingReporter();
 
   @Override public void report(@NotNull Problem problem) {
-    var render = problem.computeFullErrorMessage(DistillerOptions.informative(), false, false, 80);
+    var render = CliReporter.errorMessage(problem, DistillerOptions.informative(), false, false, 80);
     if (problem.level() != Problem.Severity.ERROR) {
       System.err.println(render);
       return;
