+++ cli/src/main/java/org/aya/cli/single/CliReporter.java
@@ -5,7 +5,9 @@ package org.aya.cli.single;
 import org.aya.api.distill.DistillerOptions;
 import org.aya.api.error.Problem;
 import org.aya.api.error.Reporter;
+import org.aya.pretty.backend.string.StringPrinterConfig;
 import org.aya.pretty.printer.PrinterConfig;
+import org.aya.pretty.style.AyaStyleFamily;
 import org.aya.util.error.SourcePos;
 import org.fusesource.jansi.AnsiConsole;
 import org.jetbrains.annotations.Contract;
@@ -38,7 +40,7 @@ public record CliReporter(
     if (level.ordinal() > minimum.ordinal()
       // If it's `SourcePos.NONE`, it's a compiler output!
       && problem.sourcePos() != SourcePos.NONE) return;
-    var errorMsg = problem.computeFullErrorMessage(options.get(), unicode.getAsBoolean(), supportAnsi, terminalWidth());
+    var errorMsg = errorMessage(problem, options.get(), unicode.getAsBoolean(), supportAnsi, terminalWidth());
     if (level == Problem.Severity.ERROR || level == Problem.Severity.WARN) err.accept(errorMsg);
     else out.accept(errorMsg);
   }
@@ -47,5 +49,18 @@ public record CliReporter(
     int w = AnsiConsole.getTerminalWidth();
     // output is redirected to a file, so it has infinite width
     return w <= 0 ? PrinterConfig.INFINITE_SIZE : w;
+  }
+
+  public static @NotNull String errorMessage(
+    @NotNull Problem problem, @NotNull DistillerOptions options,
+    boolean unicode, boolean supportAnsi, int pageWidth
+  ) {
+    var doc = problem.sourcePos() == SourcePos.NONE ? problem.describe(options) : problem.toPrettyError(options).toDoc();
+    if (supportAnsi) {
+      var config = StringPrinterConfig.unixTerminal(pageWidth, unicode);
+      config.getStylist().setStyleFamily(AyaStyleFamily.ADAPTIVE_CLI);
+      return doc.renderToString(config);
+    }
+    return doc.renderWithPageWidth(pageWidth, unicode);
   }
 }
